contract_engine/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-multipart==0.0.20
Pillow==11.1.0
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pypdf==5.3.0
PyMuPDF==1.25.3
